+++ docs/components/DocsComponent.js
@@ -1,7 +1,7 @@
 import { kebabcase } from "../../fachwerk.js";
 
 export default {
-  props: ["title", "c"],
+  props: ["title", "c", "type"],
   methods: {
     propsTable(props) {
       return Object.entries(props).map(p => ({
@@ -79,9 +79,9 @@ Component can be imported using Javascript import:
   template: `
   <f-content-editor
     type="document"
-    :hidetype="true"
-    :menu="false"
-    :save-id="title"
+    edit="show"
+    :title="title"
+    :type="type"
     :content="generateContent(title,c)"
   />
   `