+++ playground/index.js
@@ -1,22 +1,7 @@
-import { fachwerk } from "../src/fachwerk.js";
-
-const FSimpleContentEditor = {
-  data: () => ({ content: "" }),
-  mounted() {
-    console.log(this.$slots.default)
-    //this.content = this.$el.bla
-  },
-  template: `
-  <div>
-    <div id="bla" v-show="false"><slot /></div>
-    <div>{{ content }}</div>
-  </div>
-  `
-};
+import { fachwerk } from "../fachwerk.js";
 
 fachwerk({
-  components: { FSimpleContentEditor },
-  editor: "show",
-  menu: "none",
-  type: "document"
+  edit: "hide",
+  title: "playground",
+  type: "slides",
 });