+++ src/components/icons/FEditIcon.js
@@ -1,0 +1,15 @@
+export default {
+  description: `
+Editor icon.
+  
+<f-editor-icon />
+
+<p />
+  `,
+  template: `
+  <svg width="16" height="16" viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path d="M12 1L15 4L6 13L1 15L3 10L12 1Z" stroke="var(--icon-stroke)" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"/>
+    <path d="M10 3L13 6" stroke="var(--icon-stroke)" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"/>
+  </svg>
+  `
+};