+++ src/components/internal/FOverlay.js
@@ -22,7 +22,7 @@ export default {
       bottom: 0;
       background: var(--white);
       overflowY: auto;
-      box-shadow: -5px 0 10px rgba(0,0,0,0.15);
+      box-shadow: -5px 0 10px rgba(0,0,0,0.1);
     "
     :style="{
       width: currentSize