+++ src/components/layout/FFooter.js
@@ -7,21 +7,13 @@ Displays a page footer.
 <br>
   `,
   template: `
-  <div class="footer">
-    <div>
-      <p>
-        All code is licenced under
-        <a href="https://choosealicense.com/licenses/mit/" rel="licence">
-          MIT licence
-        </a>.
-        All content is licenced under
-        <br>
-        <a rel="license" href="https://creativecommons.org/licenses/by-nc-sa/4.0/">
-          Creative Commons BY-NC-SA 4.0 International License
-        </a>.
-      </p>
-    </div>
-    <img src="https://designstem.github.io/fachwerk/images/erasmus_logo.svg" style="width:240px" />
+  <div class="grid" style="--cols: auto 1fr auto">
+    <p style="opacity: 0.5">
+      All code is licenced under <a href="https://choosealicense.com/licenses/mit/" rel="licence">MIT licence</a>.<br>All content is licenced under Creative Commons
+      <a rel="license" href="https://creativecommons.org/licenses/by-nc-sa/4.0/">BY-NC-SA 4.0</a> International License.
+    </p>
+    &nbsp;
+    <img src="https://designstem.github.io/fachwerk/images/erasmus_logo.svg" style="width: 240px" />
   </div>
   `
 }